Combine the reset and rst_n signal
In the origin nand2tetris design, the register is not with rst_n.
The reset siginal rise but the register value will not clear.
I add the rst_n signal for register. And then combine to the reset
signal which is designed by nand2tetris. It is real case for physical
design by IC set.
After the combination, the origin golden for nand2tetris's golden file
will fall at the reset signal rise, becase the register is cleared but
the golden not. I think it is right behavior for my design. So I will
regard the output as pass.